Zyscrape: Log errors and success
Rather than returning a boolean to indicate if a submission contained a
search string, return a success message, and the timestamp of the
matching submission. In the case of an error, log and return the failed
student.

In the future, it may be useful to retry failed submissions in the case
of a bad connection rather than just logging the error.