Zyscrape: Remove late submissions
If an assignment has specified the option in the JSON file, remove late
submissions from the zybooks response. The due date/time is stored in
the JSON file, and each submission is compared to the time. If the
submission is after the due time then it is discarded. This means that
if a student did not submit before the due date, it will be considered
as "No Submission".