Fix: Bad Zip url
Occasionally the url for the submission zip file reported by zybooks is
bad. A bug report has been sent in to zybooks to inform them of the
issue, and they will look into it. Meanwhile, this workaround helps by
informing the grader that some files could not be downloaded and why.