Add changelog menu option
Adds a new menu option to show a changelog. Rather than storing the
versioning messages in versioning.py, they are all moved to
changelog.txt. When a versioning message is displayed, it is pulled from
the changelog.txt file. The menu option to show the whole changelog
just shows the contents of the changelog.txt file.