Grader: Add ability to pick submissions to grade
Adds a new option (experimental) to pick a student submission to grade.
To enable set PICK=YES as an environment variable before running.

This works by prompting the grader to select a part (if multiple exist)
then a list of submissions are shown. Once the submission is selected,
the Submission object is updated with that new data and the popup is
refreshed.