Fix: Run code not suspending curses
Running student code was also sending input to the input thread of zygrader.

Ensure that zygrader (curses) is suspended and block the input thread. The previous
commits are necessary to lead up to this, because input was blocking. By making input
non-blocking and using a shared queue to share input, the thread can be paused at any time