Class Manager: Update lab list while editing
Filtered lists that use a callback now pass the list itself as an argument
to the callback function.

This allows refreshing the list from other components. While it is a little
hacky to do this (removing control from the wm), zygrader is small enough that
this is simpler than designing a more robust window manager.