Grader/Window: Cleanup and fix students remaining colored red
When the feature to automatically update the list on file changes was
added, it was coded into the filtered list creation code. This was messy
because it didn't allow for much flexibility.

Now that creating a filtered list will pass the filtered list into the
callback function, the callback has the ability to force a refresh.

This adds a new window function push_refresh_event will force the window
to refresh (because the UI is multithreaded) by sending an event into the
queue. This can be called from the main thread or from other threads to
redraw.