Use waiting popup for logging into zybooks
Instead of prompting the user to press Enter after logging into zybooks,
just use a popup that automatically closes.

Now we have a waiting popup and the logger that are components that get
controlled after creation outside of window.py. Maybe we need a way to break
these off better?