Grader: Fix display of students with no submission
Students with no submission were passing through to the grader if the assignment
had an attached due date, rather than showing the "No Submission" popup.

This modifies the code a bit to always show the grader, but to show a more descriptive
message when a student has not submitted, either at all, or before the due date.