Fix: DatetimeSpinner crash on invalid dates
Resolve a few cases where the date was invalid on changing the month or year.
For example, changing to June when the date was July 31.

The solution is to pick the date that makes sense, in the above case switch to the 30th
of June.

Also handle leap year and leap day.