Fix: Grade Puller cancel add assignment
In the grade puller it prompts to "Add another assignment to the report?". If you choose yes,
it lists the assignments. However, choosing "Back" from this list exits the grade puller, not
just that list. Add a case to catch the stop and only exit the new assignment selection, not the
grade puller itself.