Cleanup: ui module interface
Begin the refactor of the window manager. A first easy step is to make it simpler
to access the window. Rather than needing to import the Window object itself,
we can simply add a function in zygrader.ui to get it for us.

From now on use ui.get_window() to access the Window