Fix: Input issues when suspending curses
After switching to nodelay, any subprocesses (less, vim, nano, etc) would not
accept input properly. This was because the input window was on a separate thread
and it would be waiting for input. The first input given to the subprocess was
consumed by the input thread, not the subprocess.

This made vim behave strangely, and the first input while running
student code needed duplication.

The solution is to switch to halfdelay, which causes curses to only wait for a few
fractions of a second before polling for input again. This allows the subprocesses
to work properly.