Preferences: Add new get() and set() functions
Simplify the usage of preferences by storing all preferences in the config
json file. Rather than relying on the presence of a key for booleans, just
store the true/false value in json because python can serialize that easily.

Use the new get() and set() functions everywhere in code, unless the json
needs to be manipulated manually (like in versioning).