Versioning: Split into two stages
This has been a small issue in the past, and it finally got to the point
where it was unavoidable.

Versioning code did two things: Modify user config for new preferences/defaults,
and show the user the changelog when a new major version was released. This caused
issues because the of a dependency between the window and the preferences.

The solution is to do preference changes, then init the window, then show
the user any changelog messages.