Window Manager Refactor
The window manager is now split into separate files.

The event handling code has been moved to events.py.
The component code has been wrapped in layers.py and is more flexible
for reuse.

There isn't a very logical way to split this into small patches, so
I am committing this in one commit so we don't flood the commit history
with 100+ commits.

Closes #39
Closes #28