Fix #13: Cleanup grade puller datetime copy
The grade puller asks to copy the date and time to all selected sections,
but this was done with two sequential popups. Use an options popup instead
to make the question more clear and require fewer actions.

This also changes the popup to always show when multiple sections are
selected, and not only when the first is selected.