Fix #78: View submissions from email manager
Add an option to the email manager to view student submissions. This requires
a small modification to the grader to optionally pass in a student. The usual
grader flow goes Lab -> Student -> Submission, but we need to skip the Student
step in this process.

So the lab list function now has an optional parameter for a student. If that
student is defined, then the Student list step is skipped and the submission is
immediately shown.

Additionally, the use_locks argument is False to prevent locking in the case
of just viewing code.

Closes #78