Remove email and password from config
Instead store the auth token in the config. Each time zygrader starts
a request is sent to see if the auth token is valid. If it isn't the
user is again prompted for email and password.

This also removes the "Save Password" user preference.
Logging out simply clears the auth token.