Fix #72: Directories created in cwd when exiting early
If zygrader would ever recieve a sigint or sighup before the shared
data was initialized it would try to remove locks. But the shared paths
were never initialized so the lock folders would be created in the current
working directory.

The solution is to only remove locks if the shared data was actually
initialized.

Closes #72