Store refresh token instead of auth token
When signing in to zyBooks an auth token and a refresh token are returned.
Previously we only stored the auth token, but it turns out this expires after
about two days from creation. After expiration the refresh token is used
to generate a new auth token.

Now we store the refresh token in the preferences. On startup we send the
refresh token to zybooks which sends us back a new auth token to use for the
session.

A better solution might be to store the expiry date as well, and then only
send a request when the auth has expired. This can be done in a follow0up
commit if needed.