Cleanup tempfiles when closing zygrader
Adds a wrapper around tempfile.TemporaryDirectory and replaces
each existing call to mkdtemp with the wrapper function. This function
stores the temporary directories in a list. When zygrader is exited
this list is destroyed, and each directory in the list is automatically
deleted.

These paths are also prefixed with "zy" to distinguish from other tmp
directories.